+++ collect_app/src/test/java/org/odk/collect/android/utilities/QRCodeUtilsTest.java
@@ -12,6 +12,8 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.preferences.AdminSharedPreferences;
 import org.odk.collect.android.preferences.GeneralSharedPreferences;
+import org.odk.collect.android.preferences.qr.ObservableQRCodeGenerator;
+import org.odk.collect.android.preferences.qr.QRCodeGenerator;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
 import org.robolectric.RobolectricTestRunner;
@@ -35,7 +37,8 @@ import static org.junit.Assert.assertTrue;
 @RunWith(RobolectricTestRunner.class)
 public class QRCodeUtilsTest {
 
-    private final File savedQrCodeImage = new File(QRCodeUtils.getQrCodeFilepath());
+    private final QRCodeGenerator qrCodeGenerator = new ObservableQRCodeGenerator();
+    private final File savedQrCodeImage = new File(qrCodeGenerator.getQrCodeFilepath());
     private final File md5File = new File(QRCodeUtils.getMd5CachePath());
 
     @Before
@@ -78,7 +81,7 @@ public class QRCodeUtilsTest {
 
         // stubbing cache and bitmap files
         new File(new StoragePathProvider().getDirPath(StorageSubdirectory.SETTINGS)).mkdirs();
-        FileUtils.saveBitmapToFile(QRCodeUtils.generateQRBitMap(expectedData, 100), QRCodeUtils.getQrCodeFilepath());
+        FileUtils.saveBitmapToFile(QRCodeUtils.generateQRBitMap(expectedData, 100), qrCodeGenerator.getQrCodeFilepath());
         FileUtils.write(md5File, getDigest(expectedData.getBytes()));
 
         // verify that QRCode and md5 cache files exist
